+++ packages/okta-auth-js/webpack.config.js
@@ -1,7 +1,7 @@
 /*
  * This config builds a minified version that can be imported
  * anywhere without any dependencies. It packages the SDK
- * with reqwest and Q. It also preserves license comments.
+ * with reqwest. It also preserves license comments.
  */
 /* global __dirname */
 var path    = require('path');