+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Breaking Changes
 
+- [#309](https://github.com/okta/okta-auth-js/pull/309) - Removed `Q` library, now using standard Promise. IE11 will require a polyfill for the `Promise` object. Use of `Promise.prototype.finally` requires Node > 10.3 for server-side use.
+
 ### Other
 
 ## 2.11.2